requirements.txt
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.0
pillow==11.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
